requirements.txt
alembic==1.13.1
alipay==0.7.4
alipay-sdk-python==3.7.33
arrow==1.3.0
attrs==23.2.0
Babel==2.14.0
bcrypt==4.1.2
Beaker==1.12.1
beautifulsoup4==4.12.3
bravado-core==6.1.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colander==2.0
coverage==7.4.4
cryptography==42.0.5
docutils==0.20.1
fqdn==1.5.1
greenlet==3.0.3
hupper==1.12.1
idna==3.6
iniconfig==2.0.0
iso8601==2.1.0
isodate==0.6.1
isoduration==20.11.0
Jinja2==3.1.3
jsonpointer==2.4
jsonref==1.1.0
jsonschema==4.21.1
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
lazy-object-proxy==1.10.0
lingua==4.15.0
Mako==1.3.2
MarkupSafe==2.1.5
more-itertools==10.2.0
msgpack==1.0.8
mysqlclient==2.2.4
openapi-core==0.19.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
optionaldict==0.1.2
packaging==24.0
parse==1.20.1
passlib==1.7.4
PasteDeploy==3.1.0
pathable==0.4.3
pillow==10.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
pluggy==1.4.0
polib==1.2.0
pyasn1==0.6.0
pycparser==2.22
pycrypto==2.6.1
pycryptodome==3.20.0
Pygments==2.17.2
PyJWT==2.8.0
pypng==0.20220715.0
pyramid==2.0.2
pyramid-jwt==1.6.1
pyramid-mako==1.1.0
pyramid-retry==2.1.1
pyramid-swagger==2.7.0
pyramid-tm==2.5
pyramid_debugtoolbar==4.12.1
pyramid_jinja2==2.10.1
pyramid_openapi3==0.18
pytest==8.1.1
pytest-cov==5.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.3
referencing==0.31.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
setuptools==69.2.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==2.0.29
swagger-spec-validator==3.0.3
transaction==4.0
translationstring==1.4
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
uri-template==1.3.0
urllib3==2.2.1
venusian==3.1.0
waitress==3.0.0
webcolors==1.13
WebOb==1.8.7
WebTest==3.0.0
wechatpay==1.1.5
wechatpayv3==1.2.50
wechatpy==1.8.18
Werkzeug==3.0.2
xmltodict==0.13.0
-e git+https://github.com/easyiit-com/ZhanorAdminPanel.git@f82b9aaed89c0775dfce5903888a42c24b60bd92#egg=zhanor_admin
zope.deprecation==5.0
zope.interface==6.2
zope.sqlalchemy==3.1
